third_party/WebKit/LayoutTests/http/tests/misc/scroll-cross-origin-iframes-expected.txt

Verify that two sibling cross-origin iframes both correctly scroll on MouseWheel events, as per https://crbug.com/675695.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS event.data.scrollTop is 40
PASS event.data.scrollTop is 40
PASS successfullyParsed is true

TEST COMPLETE
